requirements.txt
pandas==1.3.5
nltk==3.8.1
spacy==3.5.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
spacy-spanish-lemmatizer==0.7
PyStemmer==2.2.0.1
gensim==4.3.0
numpy==1.23.0
scikit_learn==1.2.1
torch==1.12.1
transformers==4.8.1
pickleshare==0.7.5